move transaction code to transaction.swift
